refactor: move JSON file loading from module
